cli.console: move progress module
