+++ src/streamlink_cli/console/progress.py
